RingFuture lifetime (#10)
This PR fixes lifetime annotations on RingFuture so that the following program will fail to compile:

use kimojio::{
    Errno,
    configuration::Configuration,
    operations::{self, OFlags},
};
async fn kimojio_main() -> Result<(), Errno> {
    let flags = OFlags::CREATE | OFlags::RDWR;
    let fd = operations::open(c"/tmp/example.txt", flags, 0o644.into()).await?;
    let buffer = b"I am ub!".to_vec();
    let fut = operations::write(&fd, &buffer);
    drop(buffer);
    let written = fut.await?;
    Ok(())
}
For RingFuture that is dropped, the implementation already prevented ub by canceling the I/O and blocking the thread until it completed. To avoid performance implications of this (not polling tasks until the canceled I/O completes) use the io_scope function.